tests/requirements.txt
pytest>=8.0
PyYAML>=6.0
paho-mqtt>=2.0
